Refactoring of how to change pages of the List.
Because the way of change page wasn't very readable and flexible.
To achieve this I have changed the HTML so that It is a unordered list (<ul>)
and all the task are list items(<li>).
The reason for using the markup ul is due to the markup div is not accessible,
and If I want to make the content reactive I need an specific markup.