Changed the name of the function saveForLS to SaveOrder
Also change the way It works. Removed the loop "for" because is better the loop
"foreach". Due to this change I see a problem with the way are created the new tasks
because It was taking the array in the reverse form, so I changed the way is inserted
new task to the application with the node.insertAdjacentElement("beforeend", element).